test: UI 테스트 코드 수정 (SeatingActivityTest 제외)
